# apps/docs/static/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/households": {
      "post": {
        "operationId": "HouseholdsController_create_v1",
        "parameters": [
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHouseholdInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Créer un foyer",
        "tags": [
          "households"
        ]
      },
      "get": {
        "operationId": "HouseholdsController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HouseholdDto"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Foyers de l'utilisateur",
        "tags": [
          "households"
        ]
      }
    },
...
    "/v1/admin/monitoring": {
      "get": {
        "description": "Onglet « Surveillance » : requêtes, erreurs, temps de réponse, sondes, incidents.",
        "operationId": "AdminController_monitoringData_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMonitoringDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Onglet « Surveillance » ",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/overview": {
      "get": {
        "description": "Administration : vue d'ensemble.",
        "operationId": "AdminController_overview_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOverviewDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Administration ",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/users": {
      "get": {
        "description": "Administration : comptes.",
        "operationId": "AdminController_users_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserDto"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Administration ",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "Administration : créer un compte.",
        "operationId": "AdminController_createUser_v1",
        "parameters": [
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreateUserInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Administration ",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/users/{id}/disable": {
      "post": {
        "description": "Administration : désactiver un compte.",
        "operationId": "AdminController_disable_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Aucun contenu"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Administration ",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/users/{id}/enable": {
      "post": {
        "description": "Administration : réactiver un compte.",
        "operationId": "AdminController_enable_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Aucun contenu"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Administration ",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/users/{id}/logout-all": {
      "post": {
        "description": "Administration : déconnecter un compte partout.",
        "operationId": "AdminController_logoutAll_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Aucun contenu"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Administration ",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/users/{id}/password-link": {
      "post": {
        "description": "Administration : lien de réinitialisation du mot de passe.",
        "operationId": "AdminController_passwordLink_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Aucun contenu"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Administration ",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/users/{id}": {
      "delete": {
        "description": "Administration : supprimer un compte.",
        "operationId": "AdminController_deleteUser_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Aucun contenu"
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Administration ",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/households": {
      "get": {
        "description": "Administration : foyers.",
        "operationId": "AdminController_households_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminHouseholdDto"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Administration ",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/backups": {
      "get": {
        "description": "Administration : sauvegardes récentes.",
        "operationId": "AdminController_backups_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BackupRunDto"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Administration ",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "Administration : lancer une sauvegarde.",
        "operationId": "AdminController_requestBackup_v1",
        "parameters": [
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupRunDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Administration ",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/test-email": {
      "post": {
        "description": "Administration : envoyer un e-mail de test.",
        "operationId": "AdminController_testEmail_v1",
        "parameters": [
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTestResultDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Administration ",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/test-push": {
      "post": {
        "description": "Administration : envoyer une notification de test.",
        "operationId": "AdminController_testPush_v1",
        "parameters": [
          {
            "name": "x-requested-with",
            "in": "header",
            "description": "Protection CSRF : valeur fixe.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "agenda-gn"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTestResultDto"
                }
              }
            }
          },
          "default": {
            "description": "Erreur (code stable dans `error.code`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Administration ",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Agenda G & N — API",
    "description": "API REST de l'application **Agenda G & N** (web et Android). Toutes les routes sont versionnées\nsous `/v1`. Guide complet (authentification, erreurs, idempotence, temps réel) : page\n*API → Guide* de la documentation.\n\n- **Authentification** : `Authorization: Bearer <accessToken>` (Android, en-tête\n  `X-Client: mobile` à la connexion pour recevoir les jetons dans le corps), ou cookies\n  `httpOnly` posés par `/v1/auth/login` (web).\n- **CSRF** : toute requête qui modifie (POST, PATCH, PUT, DELETE) doit porter\n  `x-requested-with: agenda-gn`.\n- **Erreurs** : toujours `{ \"error\": { \"code\", \"message\", \"details\"? } }` ; le `code` est\n  stable (voir le schéma `ApiError`), le `message` n'est destiné qu'aux développeurs.\n- **Idempotence** : les créations de tâches acceptent `Idempotency-Key` (UUID) pour être\n  rejouées sans doublon depuis une file hors ligne.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Même origine que le site (le web relaie /v1/* vers l’API)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gn_at"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_FAILED",
                  "UNAUTHENTICATED",
                  "INVALID_CREDENTIALS",
                  "EMAIL_ALREADY_USED",
                  "REGISTRATION_CLOSED",
                  "ACCOUNT_DISABLED",
                  "SESSION_EXPIRED",
                  "CSRF_REJECTED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "HOUSEHOLD_NOT_FOUND",
                  "INVITATION_INVALID",
                  "RESET_TOKEN_INVALID",
                  "GOOGLE_NOT_CONFIGURED",
                  "GOOGLE_FAILED",
                  "GOOGLE_EMAIL_EXISTS",
                  "GOOGLE_ALREADY_LINKED",
                  "CURRENT_PASSWORD_INVALID",
                  "PASSWORD_REQUIRED",
                  "CONFIRMATION_REQUIRED",
                  "CALENDAR_NOT_CONFIGURED",
                  "CALENDAR_NOT_CONNECTED",
                  "CALENDAR_READ_ONLY",
                  "CALENDAR_NOT_FOUND",
                  "GOOGLE_REVOKED",
                  "GOOGLE_FLOW_EXPIRED",
                  "GOOGLE_DENIED",
                  "GOOGLE_SCOPES_MISSING",
                  "ALREADY_MEMBER",
                  "VERSION_CONFLICT",
                  "IDEMPOTENCY_KEY_INVALID",
                  "IDEMPOTENCY_KEY_REUSED",
                  "IDEMPOTENCY_IN_PROGRESS",
                  "CATEGORY_NAME_TAKEN",
                  "TASK_TITLE_REQUIRED",
                  "RATE_LIMITED",
                  "ATTACHMENT_TOO_LARGE",
                  "ATTACHMENTS_QUOTA",
                  "INTERNAL"
                ]
              },
              "message": {
                "type": "string",
                "description": "Message technique (anglais), jamais affiché."
              },
              "details": {
                "description": "Détail éventuel (champs invalides pour VALIDATION_FAILED)."
              }
            }
          }
        }
      },
...
      "AdminOverviewDto": {
        "type": "object",
        "properties": {
          "counts": {
            "type": "object",
            "properties": {
              "users": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "disabledUsers": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "households": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "tasks": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "openOccurrences": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "doneLast7Days": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "shoppingItems": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "comments": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "attachments": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "users",
              "disabledUsers",
              "households",
              "tasks",
              "openOccurrences",
              "doneLast7Days",
              "shoppingItems",
              "comments",
              "attachments"
            ],
            "additionalProperties": false
          },
          "storage": {
            "type": "object",
            "properties": {
              "databaseBytes": {
                "type": "number"
              },
              "attachmentsBytes": {
                "type": "number"
              }
            },
            "required": [
              "databaseBytes",
              "attachmentsBytes"
            ],
            "additionalProperties": false
          },
          "calendar": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            }
          },
          "integrations": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": {
                  "type": "string"
                },
                "enabled": {
                  "type": "boolean"
                }
              },
              "required": [
                "key",
                "enabled"
              ],
              "additionalProperties": false
            }
          },
          "registrationEnabled": {
            "type": "boolean"
          },
          "backups": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "trigger": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "PENDING",
                    "RUNNING",
                    "OK",
                    "FAILED"
                  ]
                },
                "createdAt": {
                  "type": "string"
                },
                "startedAt": {
                  "nullable": true,
                  "type": "string"
                },
                "finishedAt": {
                  "nullable": true,
                  "type": "string"
                },
                "file": {
                  "nullable": true,
                  "type": "string"
                },
                "sizeBytes": {
                  "nullable": true,
                  "type": "number"
                },
                "summary": {
                  "nullable": true,
                  "type": "string"
                },
                "offsite": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "trigger",
                "status",
                "createdAt",
                "startedAt",
                "finishedAt",
                "file",
                "sizeBytes",
                "summary",
                "offsite"
              ],
              "additionalProperties": false
            }
          },
          "serverTime": {
            "type": "string"
          },
          "uptimeSeconds": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "counts",
          "storage",
          "calendar",
          "integrations",
          "registrationEnabled",
          "backups",
          "serverTime",
          "uptimeSeconds"
        ],
        "additionalProperties": false
      },
...
      "CreateHouseholdInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "timezone": {
            "default": "Europe/Brussels",
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "memberDisplayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          }
        },
        "required": [
          "name"
        ]
      },
...
      "HouseholdDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "userId": {
                  "nullable": true,
                  "type": "string",
                  "format": "uuid"
                },
                "displayName": {
                  "type": "string"
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "OWNER",
                    "MEMBER"
                  ]
                },
                "color": {
                  "type": "string",
                  "enum": [
                    "sage",
                    "ocean",
                    "amber",
                    "plum",
                    "clay",
                    "slate"
                  ]
                },
                "absentUntil": {
                  "nullable": true,
                  "type": "string"
                }
              },
              "required": [
                "id",
                "userId",
                "displayName",
                "role",
                "color"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "id",
          "name",
          "timezone",
          "members"
        ],
        "additionalProperties": false
      },
...
    }
  }
}